refactor: reorganize results components and rendering logic
