+++ src/App.tsx
@@ -1,11 +1,10 @@
 import { type JSX } from 'react';
 import useFetchData from './hooks/useFetchData';
+import useLocalStorage from './hooks/useLocalStorage';
 
 import Layout from '@/components/layout/Layout';
-import CardList from './components/catalog/CardList';
-import Spinner from './components/shared/ui/Spinner';
 import Search from './components/search/Search';
-import useLocalStorage from './hooks/useLocalStorage';
+import SearchResults from './components/results/SearchResults';
 
 const App = (): JSX.Element => {
   const [searchQuery, setSearchQuery] = useLocalStorage('searchQuery', '');
@@ -15,22 +14,10 @@ const App = (): JSX.Element => {
     setSearchQuery(query);
   };
 
-  const isEmpty = items.length === 0;
-
   return (
     <Layout>
       <Search onSearch={handleSearch} searchQuery={searchQuery} />
-      {isLoading ? (
-        <Spinner />
-      ) : error ? (
-        <p>{error}</p>
-      ) : isEmpty ? (
-        <p role="alert" className="text-center text-[18px]">
-          Nothing found matching &quot;{searchQuery}&quot;
-        </p>
-      ) : (
-        <CardList items={items} />
-      )}
+      <SearchResults items={items} error={error} isLoading={isLoading} searchQuery={searchQuery} />
     </Layout>
   );
 };
