+++ src/components/results/Card.tsx
