+++ tests/components/results/CardList.test.tsx
@@ -1,9 +1,9 @@
-import CardList from '@/components/catalog/CardList';
+import CardList from '@/components/results/CardList';
 import { render, screen } from '@testing-library/react';
 import { mockItems } from '../../mocks/data/items';
 import type { CardProps } from '@/types/types';
 
-jest.mock('@/components/catalog/Card', () => ({
+jest.mock('@/components/results/Card', () => ({
   __esModule: true,
   default: ({ item }: CardProps) => <div>{item.title}</div>
 }));
